# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createlabel": {
            "post": {
                "summary": "create label",
                "description": "create label",
                "operationId": "0605a13b036bfd1ed187ac2eff866df1",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "label_name"
                                ],
                                "properties": {
                                    "label_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "label created Sucessfully"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/readlabelbyid": {
            "post": {
                "summary": "Read Label",
                "description": " Read Label ",
                "operationId": "a8420b6b80b44ba893b0abd58cf4a817",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "label found Sucessfully"
                    },
                    "404": {
                        "description": "label not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/readalllabels": {
            "get": {
                "summary": "read labels",
                "description": "user read labels",
                "operationId": "5259b1cd3d277a6e0938cd079755dc77",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "labels shown suucessfully"
                    },
                    "401": {
                        "description": "No label created by this user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updatelabelbyid": {
            "post": {
                "summary": "update label",
                "description": "update user label",
                "operationId": "cfa83d29f9e06178aa931e169d1f3998",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "Updated_name",
                                    "id"
                                ],
                                "properties": {
                                    "Updated_name": {
                                        "type": "string"
                                    },
                                    "id": {}
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label successfully updated"
                    },
                    "402": {
                        "description": "labels not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deletelabelbyid": {
            "delete": {
                "summary": "delete label",
                "description": "delete user label",
                "operationId": "b3492bec79bc68cb98637639761a3046",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "label successfully deleted"
                    },
                    "404": {
                        "description": "labels not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}